feat: extract TaskRepository fetch method to UseCase layer as TaskQueryService
