+++ app/infrastructure/task/TaskRepositoryImpl.scala
@@ -7,7 +7,6 @@ package infrastructure.task
 import domain.model.task.{ Task, TaskRepository }
 import ixias.slick.SlickRepository
 import ixias.slick.jdbc.MySQLProfile.api._
-import usecase.task.ShowTaskUseCaseDto
 
 import javax.inject._
 import scala.concurrent.{ ExecutionContext, Future }
@@ -15,25 +14,9 @@ import scala.concurrent.{ ExecutionContext, Future }
 @Singleton
 class TaskRepositoryImpl @Inject() (
   @Named("master") master: Database,
-  @Named("slave") slave:   Database
 )(implicit val ec:         ExecutionContext) extends SlickRepository[Task.Id, Task] with TaskRepository {
 
-  val taskTable     = TableQuery[TaskTable]
-  val categoryTable = TableQuery[CategoryTable]
-
-  def fetchAll() = {
-    val showData = for {
-      (task, category) <- taskTable join categoryTable on (_.categoryId === _.id)
-    } yield (task.title, task.body, task.state, category.name, category.color)
-    slave.run(showData.result).map(_.map((ShowTaskUseCaseDto.apply _).tupled))
-  }
-
-  /**
-    * Get Task Data
-    */
-  def getById(id: Task.Id): Future[Option[Task]] = {
-    slave.run(taskTable.filter(_.id === id).result.headOption)
-  }
+  val taskTable = TableQuery[TaskTable]
 
   /**
     * Add Task Data
