+++ app/usecase/task/ShowTaskUseCaseDto.scala
@@ -1,5 +1,5 @@
 package usecase.task
 
-import domain.model.task.{Color, Status}
+import domain.model.task.{ Color, Status }
 
 case class ShowTaskUseCaseDto(title: String, body: String, state: Status, categoryName: String, categoryColor: Color)
