+++ app/usecase/task/TaskQueryService.scala
@@ -1,0 +1,14 @@
+package usecase.task
+
+import domain.model.task.Task
+
+import scala.concurrent.Future
+
+trait TaskQueryService {
+  def fetchAll(): Future[Seq[ShowTaskUseCaseDto]]
+
+  /**
+   * Get Task Data
+   */
+  def getById(id: Task.Id): Future[Option[Task]]
+}
